pyproject.toml
[project]
name = "tomboone-dot-com-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Tom Boone",email = "tbc@tomboone.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlmodel (>=0.0.21,<0.1.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "fastapi-azure-auth (>=5.0.0,<6.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "pymysql (>=1.1.2,<2.0.0)"
]

[tool.poetry]
packages = [{include = "app"}]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
mypy = "^1.17.1"
ruff = "^0.12.10"
bandit = "^1.8.6"
pre-commit = "^4.3.0"
pytest-cov = "^6.2.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
